requirements.txt
absl-py==2.3.1
accelerate==0.33.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
apollo-torch==1.0.3
attrs==25.3.0
bert-score==0.3.13
bitsandbytes==0.46.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
dotenv==0.9.9
evaluate==0.4.4
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec
greenlet==3.2.3
grpcio==1.73.1
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.5.1
multiprocess==0.70.16
networkx==3.5
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.2.5
nvidia-nvtx-cu12==12.8.90
optuna==4.4.0
packaging==25.0
pandas==2.3.0
peft
pillow==11.2.1
polars==1.31.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rouge-score==0.1.2
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
SQLAlchemy==2.0.41
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.43.4
triton==3.4.0
trl
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
faiss-cpu==1.7.4
gradio
